Load odoo registry before running tests
Initialize the registry before running tests.
If we don't do that, the modules will be loaded *inside* of the first
test we run, which would trigger the launch of the postinstall tests
(because we force 'test_enable' to True and the at end of the loading of
the registry, the postinstall tests are run when test_enable is enabled).
And also give wrong timing indications.