Prevent duplicate Odoo module loading.
Prior to this commit, any odoo module was loaded in duplicate.

For example, with the module report_aeroo:

First, the module was loaded as odoo.addons.report_aeroo,
then, it was loaded a second time as report_aeroo (without odoo.addons).

Therefore, any model inside the module was declared 2 times.