Mock Odoo module for unit tests
As Odoo is an heavy piece of software and we don't have to check if it runs.
So we can mock the import.
For further unit tests, the few methods we use will have to be mocked too.

Changes in the CI

replaces unittest by pytest
configure codecov
use tox

test pytest addon with pytest

Test with pytester

use pytest-cov

Use tox

fix coverage data collection

xml report for codecov

Set token for codecov

Upgrade GH action codecov to v4